requirements.txt
Django==1.10
djangorestframework==3.4.3
Pygments==2.1.3
gunicorn==19.6.0